requirements.txt
websocket-client~=1.3.2
rel~=0.4.7
pandas~=1.4.2
redis~=4.1.4
aiohttp==3.7.4.post0
uvloop==0.16.0
retrying-async==1.2.0
kafka-python==2.0.2
async-timeout==3.0.1
python-dotenv==0.19.2
orjson==3.7.5
